feat: permit omitting config var definitions
and make it possible to selectively apply the (heretofore) default
configuration.  That is, don't mandate a default `kodi_language`, don't
automatically apply `skin.estuary` settings, and so on, but instead
provide the values as a kind of policy hidden behind a boolean flag.

BREAKING CHANGE: the default configuration now omits `kodi_language`,
`kodi_locale_country`, etc., and also omits a bunch of items that
previously appeared in `kodi_config`.