feat: support dep resolution for all addons
even those that are (a) already installed or (b) installed as a
"oneshot" from a URL.  This permits removing the restriction of such
oneshot installations to `repository.*` addons; now, we attempt to
support all addon types.

This change includes support for fetching dependency information from
`~/.kodi/addons/<addon>/addon.xml`, in addition to fetching it from the
various repository `addons.xml` files.  This is because we cannot count
on those repositories to carry dependency information for oneshot addon
installations.