refactor: add "is Kodi running?" control vars
Namely, add `kodi_check_process_cmd` for specifying the command to
pass to the `cmd` argument of the `shell` module, and add
`kodi_check_process_executable` for specifying the executable to pass to
the `executable` argument of the `shell` module.

Incidental change: if `kodi_systemd_service` is defined, do not check
whether Kodi is running; instead, assume that stopping the specified
service will kill the relevant Kodi process(es).