fix: ignore errors creating `kodi_groups`
as some supported distributions (including LibreELEC) may lack
`groupadd` and/or other prerequisites of the Ansible `group` module.  If
the group-creation tasks fail and one or more target groups do not
already exist, then the task that creates `kodi_user` and adds it to
`kodi_groups` will fail (so group-creation failures are safe to ignore,
as any real problems will bubble up downstream).