fix(tests): ensure `kodi_codename` is defined
to avoid an undefined variable error on LibreELEC when expanding the
URL of the official Kodi repository URL.  Do this by (a) filtering the
special `omit` token from the `kodi_repositories` list, (b) getting rid
of the `set_fact` task that extends `kodi_repositories` on LibreELEC
only, and (c) replacing it with an entry in `kodi_repositories` that
evaluates to a repository URL on LibreELEC and `omit` on other
distributions.

Taken together, this permits deferring the evaluation of `kodi_codename`
until that variable has been initializaed by the `facts.yml` tasks file.